Use new API to create fba manifest
